remove code duplication in dictionary converters
