+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Collection/IReadOnlyDictionaryOfTKeyTValueConverter.cs
@@ -6,9 +6,9 @@ using System.Text.Json.Serialization.Metadata;
 
 namespace System.Text.Json.Serialization.Converters
 {
-    internal sealed class IReadOnlyDictionaryOfTKeyTValueConverter<TCollection, TKey, TValue>
-        : DictionaryDefaultConverter<TCollection, TKey, TValue>
-        where TCollection : IReadOnlyDictionary<TKey, TValue>
+    internal sealed class IReadOnlyDictionaryOfTKeyTValueConverter<TDictionary, TKey, TValue>
+        : DictionaryDefaultConverter<TDictionary, TKey, TValue>
+        where TDictionary : IReadOnlyDictionary<TKey, TValue>
         where TKey : notnull
     {
         protected override void Add(TKey key, in TValue value, JsonSerializerOptions options, ref ReadStack state)
@@ -24,57 +24,6 @@ namespace System.Text.Json.Serialization.Converters
             }
 
             state.Current.ReturnValue = new Dictionary<TKey, TValue>();
-        }
-
-        protected internal override bool OnWriteResume(Utf8JsonWriter writer, TCollection value, JsonSerializerOptions options, ref WriteStack state)
-        {
-            IEnumerator<KeyValuePair<TKey, TValue>> enumerator;
-            if (state.Current.CollectionEnumerator == null)
-            {
-                enumerator = value.GetEnumerator();
-                if (!enumerator.MoveNext())
-                {
-                    enumerator.Dispose();
-                    return true;
-                }
-            }
-            else
-            {
-                enumerator = (Dictionary<TKey, TValue>.Enumerator)state.Current.CollectionEnumerator;
-            }
-
-            JsonTypeInfo typeInfo = state.Current.JsonTypeInfo;
-            _keyConverter ??= GetConverter<TKey>(typeInfo.KeyTypeInfo!);
-            _valueConverter ??= GetConverter<TValue>(typeInfo.ElementTypeInfo!);
-
-            do
-            {
-                if (ShouldFlush(writer, ref state))
-                {
-                    state.Current.CollectionEnumerator = enumerator;
-                    return false;
-                }
-
-                if (state.Current.PropertyState < StackFramePropertyState.Name)
-                {
-                    state.Current.PropertyState = StackFramePropertyState.Name;
-
-                    TKey key = enumerator.Current.Key;
-                    _keyConverter.WriteWithQuotes(writer, key, options, ref state);
-                }
-
-                TValue element = enumerator.Current.Value;
-                if (!_valueConverter.TryWrite(writer, element, options, ref state))
-                {
-                    state.Current.CollectionEnumerator = enumerator;
-                    return false;
-                }
-
-                state.Current.EndDictionaryElement();
-            } while (enumerator.MoveNext());
-
-            enumerator.Dispose();
-            return true;
         }
 
         internal override Type RuntimeType => typeof(Dictionary<TKey, TValue>);
